+++ mrv2/lib/mrvGL/mrvTimelineViewportPrivate.h
@@ -55,6 +55,10 @@ namespace mrv
         std::shared_ptr<imaging::FontSystem> fontSystem;
         Fl_Menu_Button* popupMenu = nullptr;
 
+        //! Temporary help text displayed in HUD
+        static std::string helpText;
+        static float       helpTextFade;
+
         //! HUD display flags (ORed together)
         static bool hudActive;
         static HudDisplay hud;